django_API_beginner/3_login_api/profiles_project/requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cx-Oracle==7.1.3
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.8.2
pycparser==2.19
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
wincertstore==0.2
